fix: change --inflight option default value to 1
The old default value `0` means `infinity`.
Which means no back-pressure from the target MQTT broker under test.

With default set to 1, for each publisher client, only one message
is put to the send buffer each time.

In conjunction with the `--interval_of_msg` option,
the 'overrun' counter report should indicate if the system
uder test is overloaded or not.